ml-service/requirements.txt
torch>=2.6.0
numpy>=1.24.3
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
tensorboard>=2.15.0
tqdm>=4.66.0

# RAG Dependencies
langchain>=0.1.0
langchain-community>=0.0.13
langchain-anthropic>=0.1.0
faiss-cpu>=1.7.4
python-docx>=1.1.0
anthropic>=0.18.0
sentence-transformers>=2.2.2
tiktoken>=0.5.2
python-dotenv>=1.0.0
